refactored performance test, upgraded to .net 6
* updated framework to .net 6
* updated nuget packages
* using System.Text.Json instead of Newtonsoft.Json
* serializing and deserializing to/from json bytes instead of string for better parity
* moved performance tests to seperate classes
